Re-work percentile to work along a given axis
This commit re-works the linear percentile function
to also an `axis` parameter and compute the percentiles
along that. This is better aligned with NumPy's behavior
but not quite. There is room to parallelize this, but
I can revisit that when I settle on a parallelization
interface thats uniform for imgal.

Thanks @gselzer for showing me how to use `AsArray`.
This function is now as flexible as it can be
with readonly arrays :tada:.