Implement AsArray data parameter type
This commit implements `AsArray<'a, T, D'>` where
readonly arrays are used. The `AsArray` preferred
type is limited to "data" parameters only. This is
because contraining all arrays to a generic creates
problems (for example an n-dimensional dataset with
an associated 1D array). Some funtions are now
n-dimensional due to using `AsArray` as the input
type for data (for example `statistics::sum` and
`integration::midpoint` functions are now n-dimensional).