Rethink the point cloud for kd_tree
I misunderstood what a kd-tree is actually doing
and how to make it n-dimensonal. I *thought* you
build a kd-tree from the source image data and then
query that for positions or something. Yeah no. You
build the tree from a point cloud. I could make a
struct that locks in dimenions...something like:

sturct Point2D {
    x: T,
    y: T,
}

But that means a user needs to construct a bunch
of points and use this inteface. Gross. What makes
sense to me is if you already have your points in
a sensible array shape like (p, k) where p is the
point in the k dimension. So a bunch of polygon
points in 2D space would be something  like
(1000, 2). This is now the cloud.