Use the point's index instead of coordinates
Use the point's index instead of the point's
coordinates for the results vector. This is cleaner
and more efficient. Potentially there could be many
results, exploding the size of the results Vec<T>.

Instead we can just keep track the point indices
and then at the end select the points from the
point cloud and return a new mini point cloud
of results.