Use slices for KDTree search query type
Do not give the query the same lifetime as the
point cloud. This makes sense generally, and what
I did. But you can't use temporary variables
with it inside an iterator because those variables
have a shorter lifetime than the point cloud (likely
outside the iterator). Yuck! Wile not as ergonomic
to use as AsArray, using just a slice here works.