Divide total_w_pairs by 2 to account for symmetry
The total_w_pairs variable is essentiall 2(C + D)
where C and D are the concordant and disconcordant
pairs respectively. This is because I use this
equation: (Σwᵢ)² - Σwᵢ² instead of a nested for loop:

 for i in 0..weights.len() {
     for j in (i + 1)..weights.len() {
         sum += weights[i] * weights[j];
     }
 }

This is because computing the total_w_pairs this
way is faster, at the cost of the result being
symmetrical. Thus all we needed to do is divide
by 2 here.