Review and consolidate comments with misc formatting
Now that I'm more experinced in Rust a lot of these
comments are really not necessary as the code itself
is rather explanitory. For example there is no need
to comment that we are creating a view of the input
data when the variable name and associated .into()
function clearly indicate that. I also fixed a couple
of typos in the comments and lightly moved a few
lines around for logic reasons (i.e. it made sense
to me to error out first before creating a view of
data if we can). In other places I improved the
comments by adding more detail, like in the kendall
tau b module.