Fix Otsu threshold result
This is a weird one. I ported this Otsu implementation
from ImageJ-Ops/SciJava Ops. As such I took the
start and stop ranges as well. For reasons that I don't
fully understand, this is *wrong*. Its right in the Ops
framework but wrong here. There must be an assumption made
when creating the histograms that I'm not making. I discovered
this bug while updating the SACA functions. The threshold results
did not match scikit-image. This commit fixes this bug.